Rework to be oriented around Dbr
Dbr appears to be a literal interchange format. Since the original design,
I decided to move to an interface-trait architecture that is agnostic of
the way the PV data are stored. This meant that the original 'Dbr'
(subsequently renamed 'Record') was a bit opinionated. Instead, treat
Dbr as literally the information sent to/from clients in the CA protocol,
and pass these to/from the Provider implementation directly.